requirements.txt
torch
torchvision
numpy
matplotlib
pillow
streamlit
